bot-api/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
psycopg2-binary
redis
python-dotenv
apscheduler
pillow
python-telegram-bot>=20.0
aioredis
httpx
sqlalchemy
asyncpg
pydantic
